+++ sources/ElkArte/Packages/PackagesFilterIterator.php
@@ -11,7 +11,7 @@
  *
  */
 
-namespace ElkArte;
+namespace ElkArte\Packages;
 
 /**
  * Extends RecursiveFilterIterator to filter anything that looks like a package.
